refactor(docs): code dx
