+++ apps/docs/content/components/code/colors.raw.jsx
@@ -1,0 +1,14 @@
+import {Code} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex flex-wrap gap-4">
+      <Code color="default">npm install @nextui-org/react</Code>
+      <Code color="primary">npm install @nextui-org/react</Code>
+      <Code color="secondary">npm install @nextui-org/react</Code>
+      <Code color="success">npm install @nextui-org/react</Code>
+      <Code color="warning">npm install @nextui-org/react</Code>
+      <Code color="danger">npm install @nextui-org/react</Code>
+    </div>
+  );
+}
