+++ apps/docs/content/components/code/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Code} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex flex-wrap gap-4">
-      <Code color="default">npm install @nextui-org/react</Code>
-      <Code color="primary">npm install @nextui-org/react</Code>
-      <Code color="secondary">npm install @nextui-org/react</Code>
-      <Code color="success">npm install @nextui-org/react</Code>
-      <Code color="warning">npm install @nextui-org/react</Code>
-      <Code color="danger">npm install @nextui-org/react</Code>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
