+++ apps/docs/content/components/code/sizes.raw.jsx
@@ -1,0 +1,11 @@
+import {Code} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex flex-col gap-4">
+      <Code size="sm">npm install @nextui-org/react</Code>
+      <Code size="md">npm install @nextui-org/react</Code>
+      <Code size="lg">npm install @nextui-org/react</Code>
+    </div>
+  );
+}
