+++ apps/docs/content/components/code/sizes.ts
@@ -1,14 +1,4 @@
-const App = `import {Code} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex flex-col gap-4">
-      <Code size="sm">npm install @nextui-org/react</Code>
-      <Code size="md">npm install @nextui-org/react</Code>
-      <Code size="lg">npm install @nextui-org/react</Code>
-    </div>
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
