+++ apps/docs/content/components/code/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Code} from "@nextui-org/react";
+
+export default function App() {
+  return <Code>npm install @nextui-org/react</Code>;
+}
