+++ apps/docs/content/components/code/usage.ts
@@ -1,10 +1,4 @@
-const App = `import {Code} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Code>npm install @nextui-org/react</Code>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
